Move test functions.
