+++ tests/test_generate.py
@@ -3,17 +3,11 @@ import time
 import os
 from xml.etree.ElementTree import Comment
 from elifecrossref import generate
+from tests import TEST_BASE_PATH, TEST_DATA_PATH, read_file_content
 from elifecrossref.conf import raw_config, parse_raw_config
 
 
-TEST_BASE_PATH = os.path.dirname(os.path.abspath(__file__)) + os.sep
-TEST_DATA_PATH = TEST_BASE_PATH + "test_data" + os.sep
 generate.TMP_DIR = TEST_BASE_PATH + "tmp" + os.sep
-
-
-def read_file_content(file_name):
-    with open(file_name, 'rb') as open_file:
-        return open_file.read()
 
 
 class TestGenerate(unittest.TestCase):
